exynos3: libs3cjpeg: read memory configuration from kernel
This improves the flexibility (not all devices have the same FFC
resolution) and also allows optimization of memory usage.

We currently allocate 4 MB to the hardware JPEG encoder, but only
1 MB is actually needed if it's only used for encoding VGA images.
Note that the thumbnail encoding function of the kernel driver is
not used.

This patch is backward compatible with older kernels, but patching
is recommended. Cherry-pick the following commits:

	http://review.cyanogenmod.org/56791
	http://review.cyanogenmod.org/56792

Change-Id: I0e89b24e56b99e8e27f5a38d5c158b23021ed035